Renamed package to sql_autoloader
