+++ __init__.py
@@ -1,0 +1,3 @@
+"""Provides access to SQLiteConnector and PostgresConnector."""
+from sql_autoloader.postgres_connector import PostgresConnector
+from sql_autoloader.sqlite_connector import SQLiteConnector
