--- etl_components/__init__.py
@@ -1,2 +1,0 @@
-"""Main etl_components imports."""
-from etl_components.sqlite_connector import SQLiteConnector
