+++ sql_autoloader/__init__.py
@@ -1,0 +1,3 @@
+"""Main sql-autoloader imports."""
+from sql_autoloader.postgres_connector import PostgresConnector
+from sql_autoloader.sqlite_connector import SQLiteConnector
