+++ sql_autoloader/exceptions.py
