--- tests/__init__.py
@@ -1,1 +1,0 @@
-"""Init for etl_components unit tests."""
